+++ lib/commands.dart
@@ -1,0 +1,10 @@
+export 'commands/build.dart';
+export 'commands/clean.dart';
+export 'commands/config.dart';
+export 'commands/doctor.dart';
+export 'commands/diff.dart';
+export 'commands/history.dart';
+export 'commands/list_clients.dart';
+export 'commands/set_up.dart';
+export 'commands/slack_test.dart';
+export 'commands/white_label.dart';