+++ lib/scripts.dart
@@ -1,3 +1,0 @@
-export 'build.dart';
-export 'clean.dart';
-export 'list_clients.dart';